+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### NEXT
 
+- Worker: `RtpStreamSend` duplicated packets are discarded ([PR #1683](https://github.com/versatica/mediasoup/pull/1683).
+
 ### 3.19.14
 
 - Worker: Fix missing system header include, which fails in GCC 15 ([PR #1679](https://github.com/versatica/mediasoup/pull/1679), credits to @upisfree).